Sprint: Change Model Variables to Properties
Modified all Models in the DAL so that instead of generating specialist
methods to interact with variables, data stored in the models is now
accessed via a standardised set of properties.

As a general comment, any Model that houses an 'id' variable has a
GETTER ONLY. Setters are done during the construction of objects, via
the constructor only, writing to the variable directly.